Move TypeEmbedding, include Grigorii's suggestions
