+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/ClassPropertyAccess.kt
@@ -7,6 +7,7 @@ package org.jetbrains.kotlin.formver.embeddings
 
 import org.jetbrains.kotlin.formver.conversion.StmtConversionContext
 import org.jetbrains.kotlin.formver.embeddings.expression.ExpEmbedding
+import org.jetbrains.kotlin.formver.embeddings.types.TypeEmbedding
 import org.jetbrains.kotlin.formver.embeddings.expression.withNewTypeInvariants
 
 // We assume that thanks to the checks done by the Kotlin compiler, a property with a
