+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/types/TypeBuilder.kt
@@ -5,8 +5,6 @@
 
 package org.jetbrains.kotlin.formver.embeddings.types
 
-import org.jetbrains.kotlin.formver.embeddings.TypeEmbedding
-import org.jetbrains.kotlin.formver.embeddings.TypeEmbeddingFlags
 
 /**
  * Builder for a `TypeEmbedding`.
