+++ plugins/formal-verification/testData/diagnostics/good_contracts/contracts_with_receivers.fir.diag.txt
@@ -1,42 +1,41 @@
 /contracts_with_receivers.kt:(148,151): info: Generated Viper text for is2:
-method f$c$Class$is2$TF$T$c$Class(this$dispatch: Ref) returns (ret$0: Ref)
+method f$c$Class$is2$TF$T$Class(this$dispatch: Ref) returns (ret$0: Ref)
   ensures df$rt$isSubtype(df$rt$typeOf(ret$0), df$rt$boolType())
   ensures df$rt$boolFromRef(ret$0) == true ==>
-    df$rt$isSubtype(df$rt$typeOf(this$dispatch), df$rt$T$c$Impl2())
+    df$rt$isSubtype(df$rt$typeOf(this$dispatch), df$rt$c$Impl2())
 {
-  inhale df$rt$isSubtype(df$rt$typeOf(this$dispatch), df$rt$T$c$Class())
+  inhale df$rt$isSubtype(df$rt$typeOf(this$dispatch), df$rt$c$Class())
   inhale acc(p$c$Class$shared(this$dispatch), wildcard)
-  ret$0 := df$rt$boolToRef(df$rt$isSubtype(df$rt$typeOf(this$dispatch), df$rt$T$c$Impl2()))
+  ret$0 := df$rt$boolToRef(df$rt$isSubtype(df$rt$typeOf(this$dispatch), df$rt$c$Impl2()))
   goto lbl$ret$0
   label lbl$ret$0
 }
 
 /contracts_with_receivers.kt:(341,359): info: Generated Viper text for is1butWithDispatch:
-method f$c$Class$is1butWithDispatch$TF$T$c$Class$T$c$Class(this$dispatch: Ref,
-  this$extension: Ref)
+method f$c$Class$is1butWithDispatch$TF$T$Class$T$Class(this$dispatch: Ref, this$extension: Ref)
   returns (ret$0: Ref)
   ensures df$rt$isSubtype(df$rt$typeOf(ret$0), df$rt$boolType())
   ensures df$rt$boolFromRef(ret$0) == true ==>
-    df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$T$c$Impl1())
+    df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$c$Impl1())
 {
-  inhale df$rt$isSubtype(df$rt$typeOf(this$dispatch), df$rt$T$c$Class())
+  inhale df$rt$isSubtype(df$rt$typeOf(this$dispatch), df$rt$c$Class())
   inhale acc(p$c$Class$shared(this$dispatch), wildcard)
-  inhale df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$T$c$Class())
+  inhale df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$c$Class())
   inhale acc(p$c$Class$shared(this$extension), wildcard)
-  ret$0 := df$rt$boolToRef(df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$T$c$Impl1()))
+  ret$0 := df$rt$boolToRef(df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$c$Impl1()))
   goto lbl$ret$0
   label lbl$ret$0
 }
 
 /contracts_with_receivers.kt:(618,621): info: Generated Viper text for is1:
-method f$is1$TF$T$c$Class(this$extension: Ref) returns (ret$0: Ref)
+method f$is1$TF$T$Class(this$extension: Ref) returns (ret$0: Ref)
   ensures df$rt$isSubtype(df$rt$typeOf(ret$0), df$rt$boolType())
   ensures df$rt$boolFromRef(ret$0) == true ==>
-    df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$T$c$Impl1())
+    df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$c$Impl1())
 {
-  inhale df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$T$c$Class())
+  inhale df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$c$Class())
   inhale acc(p$c$Class$shared(this$extension), wildcard)
-  ret$0 := df$rt$boolToRef(df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$T$c$Impl1()))
+  ret$0 := df$rt$boolToRef(df$rt$isSubtype(df$rt$typeOf(this$extension), df$rt$c$Impl1()))
   goto lbl$ret$0
   label lbl$ret$0
 }
